requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

SQLAlchemy==2.0.36
alembic==1.14.0

pydantic==2.10.3
pydantic-settings==2.6.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.20

pytest==8.3.4
httpx==0.28.1
